requirements.txt
fastapi
uvicorn
passlib
bcrypt==4.0.1
python-jose[cryptography]
pyotp
qrcode
pillow
pydantic>=2.0